requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cleanup==7.0.0
django-cors-headers==4.0.0
django-dynamic-preferences==1.15.0
django-filter==23.2
django-pwa==1.1.0
django-restql==0.15.3
django-rql==4.3.0
django-rql-filter==0.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
grako==3.12.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lark-parser==0.11.0
lib-rql==1.1.8
MarkupSafe==2.1.2
packaging==23.1
persisting-theory==1.0
Pillow==9.5.0
promise==2.3
psycopg2==2.9.6
PyJWT==1.7.1
pyPEG2==2.15.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.0.2
